+++ doc/changelog.md
@@ -24,3 +24,6 @@
 - setExchange now takes _id, like all the other endpoints. Also clarified that all these endpoints take _id, not shortUser...
 
 - **API** change listJobsByUser from `/jobs/{sam}` to `/users/{id}/jobs`
+- **API** autosuggest returns UserQuery instead of User
+- **API** setTicket returns PutResponse instead of String
+- **API**